Re-implement Scanner Protocol
Now, this must be a class instead of a function. This is mean to hide the implementation of the initialization of each strategy.

For instance, this may be interesting to set workers amount for Pool strategy, which won't happen to Serial.

For that reason, a pure function would not be the best idea for this case.