Deprecate pool strategy
I've noticed that a lot of data was not being added to Postgres and the reason are Deadlocks.

Basically, I have two alternatives here: 1) Run everything in serial but in bulk or 2) Run each directory in parallel but with smaller chunk of data.

The problem is: I've tested it with buffer of 5, and this was pretty failable. My `failed.log` was full of data, some times this ran well, some times didn't...

So, I prefer to preserve data's integrity and simply remove Pool strategy from scanning.