Threat advisor as common field
This is guaranteeded by code that advisor exists. But, this isn't guaranteeded that he is in the database in the moment of ingestion or after.

For that reason, I'm setting it as a common field. This is not testable because of my current small test dataset.

After the ingestion, modify this field to be a foreign key.