Lazy load LoadFilters
This fix the issue of trying to initiate a fresh database. When you run `vitae` or `vitae init`, you'll get database schema error, since the models does not exists. The logic is right, they actually doesn't, that is why we are initializing them at all...

When you don't lazy load those models, you get this kind of issue. So, to avoid this error, the best is to load them only when needed.