Supress file not found issue
When running `init` command on a fresh cloned local repository, there is no `log` folder, so when trying to remove it, you'll get an exception.
To solve this, I'm supressing this exception.

I prefer to supress instead of check with a conditional because we should Ask for forgiven, not permission.