Add initials function to Jinja environment
This function adds abstraction and secures this won't break on production. This should simplify the logic of rendering the initials of the researcher securing this won't break due to some mistake, like getting the first value of an empty string.

This is what was happening on production, so this function solves this in an elegant way.